Don't pass a list of "ignored vbuckets" to janitor_agents.
Ignored vbuckets are the vbuckets that ns_janitor finds in a state
that it doesn't know how to deal with. The original intention was for
janitor_agents to skip these vbuckets when applying a bucket
config. But that has never been implemented, instead janitor_agent-s
would simply crash. In practice we've never seen this happen.

The issue is that before calling into janitor_agent-s, ns_janitor also
does other things, like deciding if vbuckets on some nodes can't be
activated because those nodes got restarted and lost data in ephemeral
buckets. But that code ignores the unsafe vbuckets altogether. Instead
of teaching this code path to deal with ignored vbuckets, I decided
that given that we've not seen such situations, it's easier to bail
out if we find such bad vbuckets. This particular patch simply crashes
in such a case, later patches will convert this condition into an
error.

Change-Id: Ief867023ea424aed1a8d6d49a4589a813a080210
Reviewed-on: http://review.couchbase.org/111421
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>