Revert "Add ns_config:ensure_persisted/0."
This reverts the following commits:

  2e504c4113ae87f86a801fb24388c904eabd01b0
  66ba406e7629b6600f5b9c5529a81fde5ea3fc87

The reason being that we decided not to handle persistence related
failure scenarios in vbucket map changes replication. We still have
other issues potentially leading to corrupt metadata that we can't
easily fix. This one is not the worst of them.

Change-Id: Ic09928024d21f207362b98081e725dd8444d7f0b
Reviewed-on: http://review.couchbase.org/111426
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>