MB-34818: MB-32881: Do not reload config when node leaves
This is partial revert for "Reload ns_config when joining and
leaving cluster"

Reloading config on leave can be dangerous because of the following
reason: while ns_config cache is being reloaded it's possible that
some other process might want to read from it.

At the same time reloading of ns_config on join seems to be safe
since we do it while ns_server is stopped.

Change-Id: I71ecee8105dac3e2828b6d5fd1b3fc52ecd33aa8
Reviewed-on: http://review.couchbase.org/111772
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>