MB-34149: Remote API to pass in master password.
This API is intended to be called by couchbase-cli to pass in the
master password. This API performs an RPC into the babysitter VM
and sets the password. It also maps the return value of set_password
function to discrete exit codes.

With our new Erlang distribution implementation, the cb_dist module
prints its progress to the stdout(which is useful to debug if anything
goes wrong). This broke the master password CLI command as the CLI
was specifically looking for strings like "ok", "retry",
"{badrpc, nodedown}" to be present in the stdout of the sub-process.
The CLI can now use the exit codes to display appropriate messages
to the user rather than scraping the stdout of the sub-process that
the CLI creates.

The following CLI change depends on this change:
http://review.couchbase.org/#/c/111938/

Change-Id: I6687c2d6fce1b143375dcb4774c45f91975857e1
Reviewed-on: http://review.couchbase.org/111937
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>