MB-35070: Indicate approx time of last suppressed log message.
When logging old suppressed messages we use the current time but this is
confusing because the message may have been from a while ago, and not
make sense anymore. Therefore, this fix includes when the last message
was relative to the timestamp of this message so that it is clear.

Change-Id: I24e18c1adf07562402be03b89e979fcf50c0ad08
Reviewed-on: http://review.couchbase.org/112110
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ajit Yagaty <ajit.yagaty@couchbase.com>