MB-30403 Display user name in doc read audit event
When a document is read via the REST API the audit event generated by
memcached displays the user as "ns_server".  The reason is that
ns_server is doing the request to memcached on behalf of the real user.

This change adds a ns_server read audit event using the "real" user
name.

In the future we'll implement "on behalf of" support between ns_server
and memcached so that the "real" user can be send with the request.

Change-Id: Ic972a5e2e9f5c463113507de43b65cd19ed21b16
Reviewed-on: http://review.couchbase.org/112371
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>