MB-35376: Export a report URI through tasks API
Right now, there is no caching of the report in the UI, but if there a
form of caching the UI doesn't have a simple way of knowing if the
cached report needs to be invalidated. We therefore provide them with an
URI which will serve to invalidate the cached report in the UI.
The URI will be changed when a new report is available.

As an added benefit, we can now fetch a specific report from the server.

Change-Id: I5378a582af5f25852e62998b2e6d37c9c19e325d
Reviewed-on: http://review.couchbase.org/112811
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>