Remove should_run_service() checks from service_stats_collector.
Because the process is only started when the service in question is
supposed to be running on this node and is shut down whenever that
changes (assuming there are no bugs in that logic in
services_stats_sup).

I believe it's fine if we get couple of intermittent errors when
should_run_service() transitions from true to false (for example,
because the node was failed over). It's possible even with the code
before this change.

Change-Id: I8086459573ccd365b43a32920d9bd098f80e69ce
Reviewed-on: http://review.couchbase.org/112999
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>