Clean up service_stats_collector's name cache periodically.
Previously it could grow without a bound because stats names that are
stored in the cache are often parameterized by the bucket or index
names. So if such bucket or index got deleted, the cached values would
never get cleaned up.

With this commit two ets tables with cached stats names are kept:
recent and stale. To check if there's a cached value, first the recent
table is checked. If nothing is found there, then the stale table is
checked. If something is found there, the name gets moved to the
recent table. If neither the recent nor the stale tables contain the
name, it's recomputed and put into the recent table. Every ten minutes
a cache rotation is initiated: the recent table becomes stale, the
stale table is cleared of all remaining values and is made
recent. This way any unused keys will get cleaned up after two
rotations.

Change-Id: I1193537f8a82990047811071146c99f044b20d15
Reviewed-on: http://review.couchbase.org/113004
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>