MB-34621: Stop displaying reason of exit in user log...
...when control connection to memcached is lost.

In the newer version of Erlang when a child exits with certain reasons
(badmatch) the EXIT message generated will also contain the stack trace.

This was causing the ns_memcached process to put out the stack trace in the
user log when one of its children died. Although we can just display the reason
and ignore the stack trace, a reason of the following kind:

{badmatch, {error, einval}}

is not very useful. So the plan currently is to indicate that the control
control connection to memcached is lost and more info can be found in debug logs.

Change-Id: I314ca91cafcaa9a651ff783c313814b9a6b3bab5
Reviewed-on: http://review.couchbase.org/113108
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>