Don't try to hibernate dcp_proxy processes.
This is useless because the process update dcp_traffic_monitor every
second, so it never stays idle long enough to trigger hibernation.

Even if dcp_traffic_monitor was out of the picture, it'd still be
pretty useless: dcp_noop requests/responses are exchanged by
memcached-s frequently enough in the default configuration for the
hibernation timeout not to ever be hit.

Change-Id: I54bb972473e0c605d07f847f572064a5dca6acc1
Reviewed-on: http://review.couchbase.org/113473
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>