MB-35578: Disallow node additions during rebalance.
We currently allow new nodes to be added while rebalance is ongoing.
This used to be ok previously but in 6.5 we try to synchronize the
config after updating the vbucket map at the end of vbucket move.
This synchronization step can fail if a new node is added to the cluster
and the config sync request is sent to the new node while it's getting
initialized.

The attempt here is to disallow node additions while rebalance is in
progress. Ideally we would like to funnel the node addition request
via the orchestrator and let orchestrator fend off this request if
it's state machine is in 'rebalancing' state. But the issue is that
the orchestrator will restart if orchestrator node gets renamed while
handling the node addition (happens when there is only one node in the
cluster) and this fails the request with HTTP 500.

To workaround this the node that handles the add-node request will
check if rebalance process is around and fails the add-node request
if rebalance process is running. This is a best effort approach.

Change-Id: I518e9aed2e20c6771ee25f411786a0e3fe9b8810
Reviewed-on: http://review.couchbase.org/113390
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>