Don't activate downstream socket until old data is processed.
Activating a socket is not necessarily cheap, especially when the
underlying socket is an ssl socket: essentially it's a call into a
process that might be busy receiving data and also might decide do
block the response until it decrypts the data to immediately send back
to the calling process. So this introduces an unnecessary delay
between when replicator process receives data from upstream memcached
and when it sends it downstream.

Change-Id: Id39bba3eeab212a33e87bd9eaecf4ab5fc9ffbbd
Reviewed-on: http://review.couchbase.org/113475
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>