MB-35560 Handle smaller failover timeouts
When the auto-failover timeout is a small value (e.g. minimum is 5
seconds) the skew between the different monitor processes provides
up to 1 second of staleness.

Here's an example.  "A" wakes up every tick and may update a value.
"B" wakes up every tick and get's "A"s currently reported value.  If
the two have ticks that are almost 1 second apart the data obtained by
"B" from "A" will be almost 1 second stale.

Time ---------|---------|---------|---------|---------|---------|
A    3         3         3         3         3         4
B            3         3         3         3         3         4

                                                           ^
                                                 timer pop |

When the timer pops the value that "B" obtained from "A" is stale.

The kv_monitor is updating bucket status once a second.  A consumer
of the status is the auto-failover mechanism.  If the timeout elapses
and auto-failover is looking at 1 second old data a auto-failover
may be initiated even though in the last second the bucket has become
warm.

Change-Id: Ic5d002c9f237b012692372a5d414da9c429315e9
Reviewed-on: http://review.couchbase.org/113533
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>