Add misc:partitionmap.
Which acts as lists:partition/2, but allows modifying the values. It's
useful in places like categorization of replies into successfull and
failed ones. Currently we tend to use lists:foldl/3 for this with
explicit state tracking. Using this function will remove some a bit of
noise.

Change-Id: Ia48defa8d2ba481457960a3d4f4a6e3994db80ef
Reviewed-on: http://review.couchbase.org/113788
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>