Exclude diverged vbuckets from the transitional vbucket map.
Transitional vbucket map is a special vbucket map that is constructed
as an intermediate step of delta recovery.

Given a current vbucket map:

[[a, undefined],
 [b, c]]

And a delta node `d' which is supposed to have vbucket 0, we construct
the vbucket map:

[[a, d],
 [b, c]]

The sole purpose of this map is to warm up the vbuckets that need to
be recovered on delta nodes. At the time when this code was written,
it seemed like a reasonable choice to reuse the already existing
bucket metadata for this purposes. The downside of this approach is
that we will also immediately create replications to all of the
vbuckets that are being recovered. Since these vbuckets might be
arbitrary behind, this means that delta nodes are potentially taking
lots of backfills and so our normal attempts not to overload nodes
during rebalance are defeated. This is especially an issue when the
replication traffic is encrypted: because erlang's TLS implementation
is very CPU hungry.

The right way to address this issue seems to be to stop abusing
vbucket maps for these purposes. Yet this is a substantial change that
is a bit risky at this point.

This patch addresses one part of the described issue. Specifically,
the vbuckets that ns_server itself explicitly deletes because they
diverged are excluded from the transitional vbucket map. So at least
these vbuckets are not going to contribute to the problem. The rest of
the vbuckets can still cause trouble, but assuming not too much data
has changed since the failover that preceded the recovery (the good
case), the issue should be less pronounced.

Change-Id: I437d6fbe9f95b3e86727458e77483adae050d657
Reviewed-on: http://review.couchbase.org/113796
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>