Get rid of ns_ssl_services_setup:supported_ciphers function ...
... to avoid confusion caused by bad name and allow filtering
of ciphers that follows in next commit

What this function was actually doing was just formatting ciphers
in the required form and setting the default ciphers to provide
backward compatibility. Both of those things should be better done
closer to the clients, in menelaus_cbauth and memcached_config_mgr.
Otherwise, it becomes harder to process/modify/filter the ciphers,
since they could be in different formats in that case.

Change-Id: I21e1b7804f7a2a7dc0e6095c311eb36bec9c7bae
Reviewed-on: http://review.couchbase.org/113864
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>