MB-33914 Allow file redaction to fail
Currently when redacting files we catch any errors (e.g. disk full),
print an error, but plow on.  The result of this is that log collection
ends up with empty files.  It's better to error out allowing the user to
remedy the situation and restart the collection.

Change-Id: I9e24f0d3ab9bbbb771bc19c9c10e21c20a5614b8
Reviewed-on: http://review.couchbase.org/114092
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Matt Carabine <matt.carabine@couchbase.com>