[bp] MB-35748 Fix rebalance stop when bucket mover exits on its own.
If rebalance stop request is sent just at the moment when one of the
bucket movers terminates, the stop request may get ignored. The issue
is in wait_for_mover function. When it receives a stop request from
the parent, it sends an exit to the mover and simply calls itself
recursively to wait till the mover exits. The problem is when the
mover terminates independently with reason 'normal'. In this case,
instead of exiting, wait_for_mover just returns 'ok', which continues
rebalance. The solution is to disentangle waiting for the mover
termination normally and when a stop request was received.

Change-Id: I2c4655bacfa0913037e567dc9cc8d5adb91eb64e
Original-Change-Id: I59bb0bcfadf8ae436f9bf371f292384b4653422a
Original-Reviewed-on: http://review.couchbase.org/114284
Original-Well-Formed: Build Bot <build@couchbase.com>
Original-Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Original-Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-on: http://review.couchbase.org/115171
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>