[bp] MB-35748 Prevent a deadlock in async termination.
The following sequence of events leads to a deadlock.

 - An async (A) is spawned. The body function sets trap_exit to true.
 - async:abort(A) is called. A's controlling process sends an exit
   signal to the executor process and waits for it to terminate.
 - Since the executor process traps exits, it continues to run and
   spawns a new async B.
 - The controlling process of B attempts to register with A and
   blocks.
 - The body of A calls async:wait(B).

Now A's controlling process waits for A's executor to terminate. A's
executor waits for B. And B waits for A's controlling process.

Similar deadlock may arise if the body of async A simply consumes an
EXIT message without actually exiting. That's what happened in
CBSE-6423 where ns_rebalancer was the process to consume the EXIT
message (fixed by one of the previous changes).

To address the issue, the controlling process continues to reply to
registration requests even when it waits for children to terminate.

Change-Id: I2f3ba2f965abdc930b4110b2d9b5b54c04e752ea
Original-Change-Id: I3c6c20d46ca7af1e7b6e226b97b2a880dbd420fd
Original-Reviewed-on: http://review.couchbase.org/114292
Original-Well-Formed: Build Bot <build@couchbase.com>
Original-Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Original-Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-on: http://review.couchbase.org/115179
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>