MB-35748 Flush {new_nodes, _} messages in leader_lease_acquirer.
Generally, we don't expect to see many {new_nodes, _} messages at the
same time. But if network is flaky and the process is slow processing
the messages, they might accumulate. So to stay on the safe side,
let's make sure we flush all of the outstanding messages and process
just the last one.

Change-Id: I34ec580617b0fc1b041e5b18d69d8a66338f0eb9
Reviewed-on: http://review.couchbase.org/114293
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>