MB-35748 Kill leader_lease_acquirer worker processes brutally.
It's suspected that due to a race condition in erlang runtime,
leader_lease_acquirer failed to terminate one of its workers in
CBSE-7275. Which lead to further unavailability of cluster operations
requiring a majority of leases.

[1] addresses the race condition. But it's also the case that there's
no need for leader_lease_acquirer to terminate the workers
gracefully. So this patch simply kills the workers brutally.

[1] https://github.com/couchbasedeps/erlang/commit/c52409a88130cc435abf1cf81499c5b0e6bd1442

Change-Id: Id2e7f640fe022ae0bbbba488d7857d6d9b1b2abf
Reviewed-on: http://review.couchbase.org/114295
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>