MB-34598 Fix accidental vbucket deletions due to node rename.
The issue may occur if node gets renamed while a janitor run is in
progress. The sequence of events is as follows:

 - ns_janitor gets a bucket config when the node name is node1.
 - node1 is renamed to node2.
 - ns_janitor sends the apply_new_config call to janitor_agent.
 - janitor_agent checks the vbucket map and finds that no vbuckets
   should reside on node2, since the vbucket map still uses the old
   name.
 - janitor_agent proceeds to delete all vbuckets.

We tried addressing this by stopping most of ns_server processes
during node rename. But that proved harder that we'd have hoped and
also introduces service unavailability.

So this patch instead works the issue around by detecting that node
rename might have happened when handling apply_new_config and
apply_new_config_replicas_phase calls in janitor_agent. If it detects
that rename occurred, it will refuse to handle the call.

Change-Id: Ia64681a60d24f8aa3428f13f39149c15b4e57725
Reviewed-on: http://review.couchbase.org/114573
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>