MB-32790 Make rebalance stop responsive.
When there are misbehaving topology aware services.

The issue is with service_rebalancer process making
service_agent:set_rebalancer() and service_agent:unset_rebalancer()
calls while trapping exits. If the service is slow to respond,
interrupting the rebalance (either manually or due to auto-failover)
can take a significant amount of time (until the timeout is hit on the
corresponding call).

The solution is to not trap exits when those potentially lengthy
operations are performed. This means that post rebalance service
cleanups will now happen asynchronously. We might want to make a
distinction between manual rebalance stop requests and those triggered
by auto-failover. But we have a precedent for such asynchrony: that's
how janitor_agent behaves. And nobody complained. We can change this
later if the need arises.

In order to prevent spurious failures in subsequent rebalance attempts
due to unfinished asynchronous cleanups,
service_agent:set_rebalancer() will from now on make sure that the
previous rebalance is cleaned up after. Old nodes will return an error
in such cases. To deal with them in the simplest possible way, a short
retry loop was added around set_rebalancer when cluster is not fully
upgraded to madhatter.

Change-Id: I99c26148893f2eb89f8c6d9d400e73266bd9f84b
Reviewed-on: http://review.couchbase.org/114776
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>