MB-34713: Deselect bucket before returning the socket...
...back to the connection pool.

Sockets from the connection pool can be fetched to perform bucket specific
operations (after issuing select_bucket command). By virtue of this,
memcached associates the socket with the bucket against which commands will
be issued. We didn't have a way to inform memcached to disassociate the
socket with the bucket as a result of which sockets could be closed by
memcached as part of its cleanup when the associated bucket got deleted.

With memcached supporting "@no bucket@" bucket name, ns_server now issues
the select_bucket command for "@no bucket@" before returning the socket
back to the connection pool. This causes the socket to be disassociated
from the bucket at the memcached level.

Change-Id: I9b6667c2ad6163ed947565fdae483e6adb1ebe46
Reviewed-on: http://review.couchbase.org/114717
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>