Don't misreport agent normal exit as a successful rebalance.
This is theoretical, but if one of the agents dies with reason normal,
the service rebalancer will also exit with reason normal. Which will
be interpreted as a successful termination of the service
rebalance. So wrap the agent's death reason to prevent this from
happening.

Change-Id: I042fcdc7b1036692b254d092c52c179d42bc2181
Reviewed-on: http://review.couchbase.org/114774
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>