MB-34713: Add 'remove_socket' API to ns_connection_pool...
...module.

This API would be used to cleanup the state maintained in the socket
pool for the socket and the process that was issued this socket. This
API must be used by the client when the socket fetched from the pool
gets closed by the other end.

Change-Id: If682eb0180092dd055604cb9a79f226fcb525c20
Reviewed-on: http://review.couchbase.org/114855
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>