MB-35594 Handle new dcp-vbtakeover status.
ep-engine will return "calculating-item-count", when it can't
confidently give an estimate of number of items left to be replicated
on a particular replication stream. This patch handles this new
status: we'll simply retry. This status should only be returned for a
short period of time after the replication is established.

We also use dcp-vbtakeover stats in ns_rebalance_observer. For initial
estimates, ns_rebalance_observer doesn't pass the connection name to
the stats call. Inside ep-engine it translated to an entirely
different code path and effectively we don't need to worry about the
statuses. For the updated estimates, ns_rebalance_observer already
expects specific statuses, so nothing needs to be changed.

Change-Id: Id79c7e2e1c704cc749bf600677ebed3c23d6f441
Reviewed-on: http://review.couchbase.org/114925
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ajit Yagaty <ajit.yagaty@couchbase.com>