MB-34713: Fetch socket from pool with select_bucket...
...command executed on it in run_check_config path.

Earlier, in the run_check_config path, when we fetched a socket from
the connection pool, select_bucket command wasn't getting executed for
the socket. This led to a bug where memcached can close socket that has
been picked up from the pool for bucket X, if the same socket had been
used to run operations for another bucket Y in the past and that bucket
Y is now getting deleted. This issue has now been fixed by running a
deselect_bucket command on the socket before adding it back to the pool.

To make use of that fix, we will now have to pass in the bucket name
while fetching a socket from the pool in the run_check_config path.

As select_bucket command would've already been run on the socket picked
up from the pool, we need to avoid running it again in the ensure_bucket
function when called from the run_check_config path.

Change-Id: I0311ac5036be7906f4ac6375aa92d4e6a61e2443
Reviewed-on: http://review.couchbase.org/114715
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>