Fix timeout diag logger process info collection.
The old code would wrap the accumulator in a list every time it
encountered a missing process.

The patch fixes it. In addition, lists:filtermap() is used instead of
lists:foldl(). The former is probably a bit more expensive, since it
needs to preserve the order of the list. But that's probably ok.

Change-Id: I8a62318071b0c4d893e3b556bc28f7e97e2b2d96
Reviewed-on: http://review.couchbase.org/115442
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>