Don't use rpc to get other node index path.
When running mixed versions with pre-MH nodes where we wanted to get the
"older" node's index path an rpc was done.  This is unneeded as the
information is in the node_storage_conf property of what is returned by
ns_doctor.

This is needed by cbas (where prior to Alice the default was the index
path) and by eventing (where prior to MH the default was the index
path).

Change-Id: I5f5ff472c4ad5c1ae706aa3f97f598c6ad4091a3
Reviewed-on: http://review.couchbase.org/115782
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>