[goport] Use child.Close() in shutdown().
Instead of childWorker.Close(). Because childWorker might not even
exist if startWorkers() panics.

The difference between Process.Close() and ProcessWorker.Close() is
that the latter also stops the stdin writer. When shutdown() is
called, it's guaranteed that either childWorker is not running (defer
p.shutdown() case), or that there are no outstanding and there will be
no writes in the future. So it's safe to use Process.Close().

Change-Id: I01bcd032e4752f43c0a94dc6b7b56a8115fba121
Reviewed-on: http://review.couchbase.org/116697
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>