MB-36606 Check wait_for_local_name results
In rename_node_in_config we call wait_for_local_name on ns_config and
ns_config_rep processes but don't check the results.  If one/both are
not running then we'll continue on and will encounter "noproc" if we try
to send a message to one of them.

Change-Id: I867da719ba7e762509ee9cc4fad8533605d6b939
Reviewed-on: http://review.couchbase.org/116891
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>