Don't check compat mode when handling rebalance_progress.
In ns_orchestrator. When the last pre-madhatter node is rebalanced
out and it polls rebalance progress, it's conceivable that
rebalance_progress call might get delivered after we switched to new
compat version. If ns_orchestrator is still in rebalancing
state (which is possible if it needs to upgrade services), then such
late call will crash it.

Change-Id: Iaf46250ee94fc80b1785297494bbbe205d0f7b15
Reviewed-on: http://review.couchbase.org/116989
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>