MB-35802 - Use new chart widgets for old stats page
This is a working prototype which can be found at
/stats_overview. This controller takes the same parameters as /stats,
so it could be a plug-in replacement.

One issue still remaining is the best way to handle the different
naming conventions. The StatsDirectory lists stats with simple names
like "swap_used", while the new chart widget wants stat names of the
form "@system.swap_used". Right now the controller does this mapping
when instantiated, but there should be a better way.

This overview does not attempt to show index stats, since they are
available on the Indexes page.

Change-Id: If28db12ac0d999b56d93891fff76af9d8f9020c0
Reviewed-on: http://review.couchbase.org/115537
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>