MB-36754 Use indexer provided stats in calculations
Use indexer provided "data_size_on_disk" and "log_space_on_disk" to
calculate fragmentation.

The frag_percent refers to external fragmentation.

For each index, plasma (Plasma is the storage that GSI uses as a part of
"Standard Global Secondary" storage mode) stores the compressed version
of in-memory data on disk. Additionally, it stores some checkpoint files
on disk. The size of the compressed data + size of checkpoint files
constitutes the "disk_size" of an index.

Plasma stores the data on disk in the form of "log structured storage".
Due to this, if any of the pages are compacted, the compacted page will
be added as a new page at the tail of the log and the old page will be
discarded. lss_data_size refers to the size of valid data i.e. the size
of data in pages that won't be discarded yet.

A component called lss_cleaner runs inside plasma storage to identify
pages, compact them and mark them invalid. log_space_on_disk refers to
all the pages on disk that are not marked for cleaning (or) pages on
disk that will be considered for next iteration of cleaning.

Change-Id: Ic919a97e784930f9a0e828ab3c62e114c3bdd1ca
Reviewed-on: http://review.couchbase.org/118073
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>