Don't allow mixing follower and leader quorums.
Here a leader quorum is a quorum that the leader node requests for an
activity it starts locally (that is, on the leader itself) and the
activity performs some sort of cluster wide operation. Most commonly a
majority quorum is requested for such activities.

The follower quorum is a special quorum value (which possibly should
have a better name) that the leader uses when it spawns activities on
remote nodes. This special quorum simply means that the activity
requires that the leader continues to maintain its lease on the remote
node in question.

Mixing leader quorums and follower quorums doesn't make sense and this
commit makes this explicit.

Change-Id: I75331fc05378044aa9cc18da41c5433ef662a318
Reviewed-on: http://review.couchbase.org/118092
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>