MB-35437 Don't terminate all activities on losing local lease.
Instead, only activities that requested a 'follower' quorum are
terminated.

Previously, if the leader lost a lease from itself, this would
immediately terminate all running leader activities. The reasoning
behind this was that if the leader can't even continuously hold its
own lease, it has no business running cluster wide activities.

This change revokes this special status. As long as an activity
continues to have the requested quorum, it won't get terminated.

This is more of a band-aid fix for the referred ticket. The system
that the test is being run on appears to be overloaded and
occasionally the leader would fail to renew some of the leases. As
long as those unlucky leases are not local, everything continues to
work. But if the local lease is the unlucky one, all leader activities
would get terminated.

Change-Id: I41257793a9a9e4b1493559681c2451cfa0da6028
Reviewed-on: http://review.couchbase.org/118093
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>