MB-36951: Reply to rebalance client after updating status
Previously we replied to the rebalance client before cleaning up
after the rebalance and in particular before resetting the rebalance
status. This means that the user can call the synchronous
controller/failOver REST API and receive a 200 in response and then
fail to be able to add a node via the orchestrator because the
rebalance_status is still set to running. (As happened in MB-35951.)

This patch changes the logic to reply to the client after the rebalance
has been cleaned up and the rebalance observer has been torn down.
Of course it means that it's possible that another rebalance could
have begun and ended after the original rebalance completed and
before the client received notification. But this was in theory
possible anyway.

The reason for fixing this is it doesn't seem like this poposed patch
has a negative effect. The reason for not fixing it is that perhaps we
should file an improvement to not have to prevent add-node during
rebalance.

Change-Id: I3576a3b8128b13e5fc18f230deed9efbd75eaf45
Reviewed-on: http://review.couchbase.org/118183
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>