MB-37039 Initialize in_flight_backfills_per_node correctly.
We changed the way we decide if a node is considered to have a
backfill for a particular move. Yet the initialization code assumed
the old definition. So in some corner cases the initial value wouldn't
have all nodes that would be involved in backfills. That would lead to
a crash when we tried to increment a backfill counter for such node.

This whole module needs to be cleaned up quite a bit, but for now I
just made a minimal change to fix the issue at hand.

Change-Id: I38bcc262ba7ccf4120465e82adde15f73e74e3be
Reviewed-on: http://review.couchbase.org/118451
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>