MB-37077 Drop support for multi-node diags.
Since it puts an undue memory burden on the node that's collecting the
information: it needs to keep in memory diagnostics from all nodes,
where each node's diagnostics can be multi-gigabyte.

It's especially malicious that we'd enable multi-node diagnostics
collection when the user asked to collect diagnostics from just one
node. Which they might want to do exactly because they want to limit
the effect of the collection on the system.

Implementation note. This change preserves the structure of
diag_handler module even though it could be reworked/simplified
significantly now that we don't need to handle remote
collections. This is because I wanted to minimize the scope of the
change.

Change-Id: Ib31b10f86e577711554fd1bfd35432698934294a
Reviewed-on: http://review.couchbase.org/118590
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>