MB-37273 Report errors uniformly in misc:raw_read_file().
Previously, if an error was encountered at file:open(), it would get
returned normally. But any error encountered during actual reading of
the file would get raised via erlang:error(). With this patch we'll
always return an error term.

The old approach makes it harder and more error prone to deal with
errors. The referred issue is an example of this:
ns_heart:grab_procfs_files() expected error returns, but didn't expect
raised errors. The result was continuous crashes on systems where
support for /proc/pressure/* files was disabled: the kernel would let
us open the files but then would return an error upon an attempt to
read from them.

Change-Id: I2228d11f546b408dc3d70ce2902ba55234271add
Reviewed-on: http://review.couchbase.org/119458
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>